==> r-scripts/00-preamble.R <==
# R packages for projects
library(tidyverse)
library(devtools)
library(tibble)
#library(rWCVP)
#library(rWCVPdata)
library(readxl)
library(readr)
library(lme4)
library(mgcv)
library(glmmTMB)
library(effects)
library(data.table)
library(DHARMa)
library(performance)
library(MetBrewer)
library(colorspace)
library(ggpattern)
library(matrixStats)
# load fonts
library(hrbrthemes)
library(sysfonts)
library(showtext)
font_paths("C:/Windows/Fonts")
font_add("Arial Narrow", regular = "arialn.ttf", italic = "arialni.ttf", 
         bold = "arialnb.ttf")
showtext_auto()
font_add_google("Roboto Condensed", "roboto")
showtext_auto()


library(conflicted)
conflict_prefer("select", "dplyr")
conflict_prefer("filter", "dplyr")
conflict_prefer("lag", "dplyr")
conflict_prefer("filter", "dplyr")
conflict_prefer("count", "dplyr")

#install.packages(c("sf", "rnaturalearth", "rnaturalearthdata"))
library(rnaturalearth)
library(rnaturalearthdata)
library("ggplot2")
library(scales)
library(ggrepel)
library(purrr)
library(ggpubr)
library(plotly)
library(maps)
library(sf)
library(rnaturalearthhires)
library("brms")
library("bayesplot") 
library("performance")
library(betareg)
library(statmod)
library(ggbeeswarm)
library(marginaleffects)
library(terra)
library(flora)
library(rWCVP)
#install.packages("rWCVPdata", repos=c("https://matildabrown.github.io/drat", "https://cloud.r-project.org"))
library(bestNormalize)
library(forcats)
library(emmeans)
library(ggsignif)
library(lmerTest)
library(ggeffects)
library(writexl)
library(MuMIn)
library(robustlmm)
library(mgcv)
library(gamm4)
library(hexbin)
library(viridis)
library(MoMAColors)
library(scico)
library(ggthemes)
library("ggsci")
library(gghalves)
library(paletteer)
library(NatParksPalettes)
library(patchwork)
font_add_google("Roboto Condensed", "roboto_condensed")
showtext_auto()
library(broom.mixed)
library(dichromat)
library(cowplot)
